Do not panic when evacuating a running lrp that doesn't exist anymore
[#136153749]

+++ handlers/evacuation_handler.go
@@ -154,6 +154,8 @@ func (h *EvacuationHandler) EvacuateCrashedActualLRP(logger lager.Logger, w http
 
 func (h *EvacuationHandler) EvacuateRunningActualLRP(logger lager.Logger, w http.ResponseWriter, req *http.Request) {
 	logger = logger.Session("evacuate-running-actual-lrp")
+	logger.Info("starting")
+	defer logger.Info("completed")
 
 	response := &models.EvacuationResponse{}
 	response.KeepContainer = true
@@ -233,6 +235,13 @@ func (h *EvacuationHandler) EvacuateRunningActualLRP(logger lager.Logger, w http
 		(instance.State == models.ActualLRPStateRunning && !instance.ActualLRPInstanceKey.Equal(request.ActualLrpInstanceKey)) ||
 		instance.State == models.ActualLRPStateCrashed {
 		response.KeepContainer = false
+
+		// if there is not evacuating instance, it probably got removed when the
+		// new instance transitioned to a Running state
+		if evacuating == nil {
+			return
+		}
+
 		err = h.db.RemoveEvacuatingActualLRP(logger, &evacuating.ActualLRPKey, &evacuating.ActualLRPInstanceKey)
 		if err == nil {
 			go h.actualHub.Emit(models.NewActualLRPRemovedEvent(&models.ActualLRPGroup{Evacuating: evacuating}))
